split js into several files
