+++ static/html/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>thing</title>
     <link rel="stylesheet" href="static/css/style.css">
-    <script src="/static/js/main.js"></script>
+    <script type="module" src="/static/js/main.js"></script>
 </head>
 
 <body>
@@ -23,7 +23,7 @@
 
   <div>
   <h3>Add Activity</h3>
-  <form id="addActivity" method="POST">
+  <form id="addActivity">
     <input type="text" id="Name" name="Name" placeholder="Activity">
     <br>  
     <label>health</label>
